+++ hmx2/constants/assets.py
@@ -1,0 +1,70 @@
+# ------ Assets ------
+ASSET_ETH = "ETH"
+ASSET_BTC = "BTC"
+ASSET_AAPL = "AAPL"
+ASSET_JPY = "JPY"
+ASSET_XAU = "XAU"
+ASSET_AMZN = "AMZN"
+ASSET_MSFT = "MSFT"
+ASSET_TSLA = "TSLA"
+ASSET_EUR = "EUR"
+ASSET_XAG = "XAG"
+ASSET_AUD = "AUD"
+ASSET_GBP = "GBP"
+ASSET_ADA = "ADA"
+ASSET_MATIC = "MATIC"
+ASSET_SUI = "SUI"
+ASSET_ARB = "ARB"
+ASSET_OP = "OP"
+ASSET_LTC = "LTC"
+ASSET_COIN = "COIN"
+ASSET_GOOG = "GOOG"
+ASSET_BNB = "BNB"
+ASSET_SOL = "SOL"
+ASSET_QQQ = "QQQ"
+ASSET_XRP = "XRP"
+ASSET_USDC = "USDC"
+ASSET_USDT = "USDT"
+ASSET_DAI = "DAI"
+ASSET_GLP = "GLP"
+ASSET_NVDA = "NVDA"
+ASSET_LINK = "LINK"
+ASSET_CHF = "CHF"
+ASSET_DOGE = "DOGE"
+ASSET_CAD = "CAD"
+ASSET_SGD = "SGD"
+ASSET_CNH = "CNH"
+ASSET_wstETH = "wstETH"
+ASSET_HKD = "HKD"
+ASSET_BCH = "BCH"
+ASSET_MEME = "MEME"
+ASSET_gmBTC = "gmBTC"
+ASSET_gmETH = "gmETH"
+ASSET_SEK = "SEK"
+ASSET_DIX = "DIX"
+ASSET_JTO = "JTO"
+ASSET_STX = "STX"
+ASSET_ORDI = "ORDI"
+ASSET_TIA = "TIA"
+ASSET_AVAX = "AVAX"
+ASSET_INJ = "INJ"
+ASSET_DOT = "DOT"
+ASSET_SEI = "SEI"
+ASSET_ATOM = "ATOM"
+ASSET_1000SHIB = "1000SHIB"
+ASSET_1000PEPE = "1000PEPE"
+ASSET_ICP = "ICP"
+ASSET_MANTA = "MANTA"
+ASSET_STRK = "STRK"
+ASSET_PYTH = "PYTH"
+
+ASSETS = [ASSET_ETH, ASSET_BTC, ASSET_AAPL, ASSET_JPY, ASSET_XAU, ASSET_AMZN,
+          ASSET_MSFT, ASSET_TSLA, ASSET_EUR, ASSET_XAG, ASSET_AUD, ASSET_GBP,
+          ASSET_ADA, ASSET_MATIC, ASSET_SUI, ASSET_ARB, ASSET_OP, ASSET_LTC,
+          ASSET_COIN, ASSET_GOOG, ASSET_BNB, ASSET_SOL, ASSET_QQQ, ASSET_XRP,
+          ASSET_USDC, ASSET_USDT, ASSET_DAI, ASSET_GLP, ASSET_NVDA, ASSET_LINK,
+          ASSET_CHF, ASSET_DOGE, ASSET_CAD, ASSET_SGD, ASSET_CNH, ASSET_wstETH,
+          ASSET_HKD, ASSET_BCH, ASSET_MEME, ASSET_gmBTC, ASSET_gmETH, ASSET_SEK,
+          ASSET_DIX, ASSET_JTO, ASSET_STX, ASSET_ORDI, ASSET_TIA, ASSET_AVAX,
+          ASSET_INJ, ASSET_DOT, ASSET_SEI, ASSET_ATOM, ASSET_1000SHIB, ASSET_1000PEPE,
+          ASSET_ICP, ASSET_MANTA, ASSET_STRK, ASSET_PYTH]
