+++ hmx2/constants/common.py
@@ -1,0 +1,15 @@
+# Address
+ADDRESS_ZERO = "0x0000000000000000000000000000000000000000"
+BYTE_ZERO = "0x0000000000000000000000000000000000000000000000000000000000000000"
+
+# Math
+MAX_UINT = 2 ** 256 - 1
+BPS = 10000
+
+EXECUTION_FEE = 3 * 10 ** 14
+
+SECONDS = 1
+MINUTES = 60
+HOURS = 3600
+DAYS = 86400
+YEARS = 31536000
