+++ hmx2/constants/contracts.py
@@ -1,0 +1,33 @@
+
+CONTRACT_ADDRESS = {
+  # Arbitrum One
+  42161: {
+    "MULTICALL_ADDRESS": "0xcA11bde05977b3631167028862bE2a173976CA11",
+    "CROSS_MARGIN_HANDLER_ADDRESS": "0xB189532c581afB4Fbe69aF6dC3CD36769525d446",
+    "LIMIT_TRADE_HANDLER_ADDRESS": "0xeE116128b9AAAdBcd1f7C18608C5114f594cf5D6",
+    "GLP_MANAGER_ADDRESS": "0x3963FfC9dff443c2A94f21b129D429891E32ec18",
+    "CONFIG_STORAGE_ADDRESS": "0xF4F7123fFe42c4C90A4bCDD2317D397E0B7d7cc0",
+    "PERP_STORAGE_ADDRESS": "0x97e94BdA44a2Df784Ab6535aaE2D62EFC6D2e303",
+    "VAULT_STORAGE_ADDRESS": "0x56CC5A9c0788e674f17F7555dC8D3e2F1C0313C0",
+    "DIX_PRICE_ADAPTER_ADDRESS": "0x222918d230c5A29F334fFb3020aD57b8CeBD1B82",
+    "GM_BTC_PRICE_ADAPTER_ADDRESS": "0x85680bba8a94c9be1DDd7Be802885DFCe95F8164",
+    "GM_ETH_PRICE_ADAPTER_ADDRESS": "0x700083c72eBc86CbFc865830F5706a2DbC392f26",
+    "TRADE_HELPER_ADDRESS": "0x963Cbe4cFcDC58795869be74b80A328b022DE00C",
+    "ONCHAIN_PRICELENS_ADDRESS": "0x7D8eAa8dF02526c711F4ff1f97F6c5324212DBBa",
+    "CALCULATOR_ADDRESS": "0x0FdE910552977041Dc8c7ef652b5a07B40B9e006",
+  }
+}
+
+# ------ ABI Path ------
+ERC20_ABI_PATH = "abis/ERC20.json"
+CROSS_MARGIN_HANDLER_ABI_PATH = "abis/CrossMarginHandler.json"
+LIMIT_TRADE_HANDLER_ABI_PATH = "abis/LimitTradeHandler.json"
+VAULT_STORAGE_ABI_PATH = "abis/VaultStorage.json"
+GLP_MANAGER_ABI_PATH = "abis/GlpManager.json"
+PERP_STORAGE_ABI_PATH = "abis/PerpStorage.json"
+CONFIG_STORAGE_ABI_PATH = "abis/ConfigStorage.json"
+CIX_PRICE_ADAPTER_ABI_PATH = "abis/CIXPriceAdapter.json"
+GM_PRICE_ADAPTER_ABI_PATH = "abis/GMPriceAdapter.json"
+TRADE_HELPER_ABI_PATH = "abis/TradeHelper.json"
+ONCHAIN_PRICELENS_ABI_PATH = "abis/OnchainPricelens.json"
+CALCULATOR_ABI_PATH = "abis/Calculator.json"
