+++ hmx2/modules/oracle/gm_oracle/gm_oracle.py
@@ -2,7 +2,7 @@ from typing import List
 from web3 import Web3
 from hmx2.helpers.contract_loader import load_contract
 from hmx2.modules.oracle.pyth_oracle import PythOracle
-from hmx2.constants import GM_PRICE_ADAPTER_ABI_PATH
+from hmx2.constants.contracts import GM_PRICE_ADAPTER_ABI_PATH
 import math
