+++ tests/helpers/pyth_price_service_helper.py
@@ -1,5 +1,5 @@
 import responses
-from hmx2.constants import DEFAULT_PYTH_PRICE_SERVICE_URL
+from hmx2.constants.pricefeed import DEFAULT_PYTH_PRICE_SERVICE_URL
 from hmx2.modules.oracle.pyth_oracle.constants import GET_LATEST_PRICE_FEEDS_ENDPOINT
 from hmx2.modules.oracle.pyth_oracle.constants import PRICE_FEED_IDS
 from urllib.parse import urljoin
